Rename utils module to avoid conflicts
