+++ features.py
@@ -3,8 +3,8 @@ import pandas as pd
 import numpy as np
 
 from memory import reduce_mem_usage
-from utils import add_missing_flags
-import utils
+from ac_utils import add_missing_flags
+import ac_utils as utils
 
 
 def clean_features(
